refactored test loaders to support new loaders
