+++ src/Codeception/TestCase/Loader/Loader.php
@@ -1,0 +1,11 @@
+<?php
+namespace Codeception\TestCase\Loader;
+
+interface Loader
+{
+    public function loadTests($filename);
+
+    public function getTests();
+
+    public function getPattern();
+}
